requirements.txt
streamlit
langchain
langchain_google_genai
dotenv
PyPDF2
python-docx